Рефакторит утилиты
